+++ R/run_app.R
@@ -9,19 +9,42 @@
 #' @examples
 #' run_app("rule_eval")
 run_app <- function(app, launch_browser = TRUE) {
-  available_apps <- list.files(system.file("shiny-apps", package = "Rnssp"))
-
-  available_apps_msg <-  available_apps %>%
-    paste0("{.var ", ., "}", collapse = "', ") %>%
-    paste0("Available apps are: ", .)
-
-  if (missing(app) || !nzchar(app) || !app %in% available_apps) {
-    available_apps_msg %>%
-      paste0("Please execute {.fn run_app} with a valid app name as an argument.\n", .) %>%
-      cli::cli_abort()
+  apps_dir <- system.file("shiny-apps", package = "Rnssp")
+  local_apps <- list.files(apps_dir)
+  remote_apps <- Rnssp::list_apps()
+  available_apps <- unique(c(local_apps, remote_apps))
+  
+  available_apps_msg <- paste0(
+    "Available apps are: ",
+    paste0("{.var ", available_apps, "}", collapse = "', ")
+  )
+  
+  if(missing(app) || !nzchar(app) || !app %in% available_apps) {
+    cli::cli_abort(
+      paste0(
+        "Please execute {.fn run_app} with a valid app name as an argument.\n", 
+        available_apps_msg
+      )
+    )
   }
-
-  app_dir <- system.file("shiny-apps", app, package = "Rnssp")
-  shiny::runApp(app_dir, display.mode = "normal", launch.browser = launch_browser) %>%
-    suppressWarnings()
+  
+  if(app %in% local_apps){
+    app_dir <- system.file("shiny-apps", app, package = "Rnssp")
+    suppressWarnings(
+      shiny::runApp(
+        app_dir, 
+        display.mode = "normal", 
+        launch.browser = launch_browser
+      )
+    )
+  } else {
+    shiny::runGitHub(
+      repo = "Rnssp-shiny-apps", 
+      username = "cdcgov",
+      subdir = app, 
+      destdir = apps_dir, 
+      ref = "master", 
+      launch.browser = launch_browser
+    )
+  }
 }