+++ R/utility.R
@@ -238,18 +238,18 @@ change_dates <- function(url, start_date = NULL, end_date = NULL) {
       str_trim()
   }
   new_startd <- ifelse(nchar(new_start) > 7,
-    as.Date(new_start, "%e%b%Y"),
-    as.Date(new_start, "%e%b%y")
+                       as.Date(new_start, "%e%b%Y"),
+                       as.Date(new_start, "%e%b%y")
   )
   new_endd <- ifelse(nchar(new_end) > 7,
-    as.Date(new_end, "%e%b%Y"),
-    as.Date(new_end, "%e%b%y")
+                     as.Date(new_end, "%e%b%Y"),
+                     as.Date(new_end, "%e%b%y")
   )
   if (new_startd > new_endd) {
     cli::cli_abort("Start Date {.field {new_start}} is posterior to End Date {.field {new_end}}.")
   }
-  str_replace(url, old_end, new_end) %>%
-    str_replace(., old_start, new_start) %>%
+  str_replace(url, paste0(prefixes[["epref"]], old_end), paste0(prefixes[["epref"]], new_end)) %>%
+    str_replace(., paste0(prefixes[["spref"]], old_start), paste0(prefixes[["spref"]], new_start)) %>%
     gsub("[[:space:]]", "", .)
 }
 
@@ -320,5 +320,48 @@ list_templates <- function(as.table = FALSE) {
     )
   } else {
     templates
+  }
+}
+
+#' List Available Shinyapps
+#'
+#' List available NSSP Shinyapps from the Rnssp-shiny-apps Github repository.
+#'
+#' @param as.table a logical, if TRUE, a data frame is returned.
+#'      Otherwise, a vector is returned (Default is FALSE).
+#'
+#' @return A data frame or a vector
+#' @seealso \href{https://github.com/CDCgov/Rnssp-shiny-apps}{https://github.com/CDCgov/Rnssp-rmd-templates}
+#' @export
+#'
+#' @examples
+#' list_apps()
+#' list_apps(as.table = TRUE)
+list_apps <- function(as.table = FALSE) {
+  req <- httr::GET("https://api.github.com/repos/cdcgov/Rnssp-shiny-apps/git/trees/master?recursive=1")
+  repoURL <- "https://raw.githubusercontent.com/cdcgov/Rnssp-shiny-apps/master"
+  httr::stop_for_status(req)
+  filelist <- unlist(lapply(httr::content(req)$tree, "[", "path"), use.names = F)
+  apps <- unique(dirname(filelist[grepl("/app.yaml$", filelist)]))
+  if (as.table) {
+    do.call(
+      rbind.data.frame,
+      lapply(
+        apps,
+        function(app) {
+          tibble::add_column(
+            tibble::as_tibble(
+              yaml::read_yaml(
+                file.path(repoURL, app, "app.yaml")
+              )
+            ),
+            .before = 1,
+            id = app
+          )
+        }
+      )
+    )
+  } else {
+    apps
   }
 }