+++ NEWS.md
@@ -1,3 +1,13 @@
+# Rnssp 0.2.1
+## Updates
+* The deprecated `$get_api_tsgraph()` method has been removed.
+* The `get_api_tsgraph` function has been removed as a consequence of the removal of the deprecated `$get_api_tsgraph` method.
+
+## Bug Fixes
+* The `Credentials` class has been updated. It now interfaces with API services no authorization requirement.
+* The `get_essence_data()` function has been upgraded to accommodate the table builder ESSENCE API.
+* The `change_dates()` function has been upgraded to fix flip-flopping start and end date values ([#13](https://github.com/CDCgov/Rnssp/issues/13))
+
 # Rnssp 0.2.0
 
 ## New Features